fix(justfile): increase sample size and reduce concurrent reqs
The llama-cpp.yml config limit the number of parallel process requests
to 4 (2048 * 4 = 8192). We can send to the llama.cpp server more that 4
reqs but if the number is too high (e.g. 32) the last send reqs could
trigger the Timeout Exception. The timeout countdown for the 32-th req
start when the first req is process.